# openapi.yaml
openapi: 3.1.0
info:
  title: API de Notas Pro
  description: Un sistema de gestión de notas desarrollado bajo estándares SDD.
  version: 1.0.0
  contact:
    name: Daniel
    email: daniel@tuempresa.com

servers:
  - url: http://localhost:8000
    description: Servidor de desarrollo local

security:
  - OAuth2PasswordBearer: []

paths:
  /auth/register:
    post:
      tags:
        - Autenticación
      summary: Registrar un nuevo usuario
      operationId: registrarUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: Usuario registrado con éxito.
        '400':
          description: El usuario ya existe o datos inválidos.

  /auth/token:
    post:
      tags:
        - Autenticación
      summary: Obtener token de acceso
      operationId: loginParaToken
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Token obtenido con éxito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Token'
        '401':
          description: Credenciales incorrectas.

  /entries:
    get:
      tags:
        - Entries
      summary: Obtener todas las entradas (Notas y Tareas)
      description: Retorna una lista de las entradas del usuario autenticado con soporte para filtrado y ordenamiento.
      operationId: listarEntries
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: type
          in: query
          schema:
            type: string
            enum: [note, task]
        - name: status_column
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high]
        - name: is_encrypted
          in: query
          schema:
            type: boolean
        - name: folder_id
          in: query
          schema:
            type: integer
        - name: is_deleted
          in: query
          schema:
            type: boolean
            default: false
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [fecha_creacion, due_date, priority, titulo]
            default: fecha_creacion
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: q
          in: query
          description: "Keyword search in title and content"
          schema:
            type: string
      responses:
        '200':
          description: Una lista de entradas obtenida con éxito.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entry'
        '401':
          description: No autorizado.
    post:
      tags:
        - Entries
      summary: Crear una nueva entrada
      operationId: crearEntry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Entry'
      responses:
        '201':
          description: Entrada creada con éxito.
        '400':
          description: Datos de entrada inválidos.
  
  /entries/{id}:
    get:
      tags:
        - Entries
      summary: Obtener una entrada específica
      operationId: obtenerEntryPorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        '404':
          description: No encontrado.
    patch:
      tags:
        - Entries
      summary: Actualizar una entrada específica (metadata o contenido)
      operationId: actualizarEntry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntryUpdate'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
        '404':
          description: No encontrado.
    delete:
      tags:
        - Entries
      summary: Eliminar permanentemente una entrada (solo si está en la papelera)
      operationId: eliminarEntry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Eliminado permanentemente.
        '400':
          description: Solo se pueden eliminar permanentemente elementos de la papelera.

  /entries/trash:
    get:
      tags:
        - Entries
      summary: Listar elementos de la papelera
      operationId: listarTrash
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Entry'

  /entries/{id}/trash:
    post:
      tags:
        - Entries
      summary: Mover una entrada a la papelera (soft delete)
      operationId: trashEntry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'

  /entries/{id}/restore:
    post:
      tags:
        - Entries
      summary: Restaurar una entrada de la papelera
      operationId: restoreEntry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'

  /assets:
    post:
      tags:
        - Assets
      summary: Subir un recurso multimedia (imagen/dibujo)
      operationId: subirAsset
      parameters:
        - name: entry_id
          in: query
          schema:
            type: integer
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'

  /assets/{asset_id}:
    get:
      tags:
        - Assets
      summary: Descargar un recurso multimedia
      operationId: obtenerAsset
      parameters:
        - name: asset_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Recurso obtenido.
        '404':
          description: No encontrado.
    delete:
      tags:
        - Assets
      summary: Eliminar un recurso multimedia
      operationId: eliminarAsset
      parameters:
        - name: asset_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Eliminado.

  /columns:
    get:
      tags:
        - Kanban
      summary: Obtener todas las columnas del tablero Kanban
      operationId: listarColumnas
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KanbanColumn'
    post:
      tags:
        - Kanban
      summary: Crear una nueva columna Kanban
      operationId: crearColumna
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KanbanColumnCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KanbanColumn'

  /columns/{id}:
    delete:
      tags:
        - Kanban
      summary: Eliminar una columna Kanban
      operationId: eliminarColumna
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Eliminado.

  /folders:
    get:
      tags:
        - Organization
      summary: Obtener todas las carpetas
      operationId: listarCarpetas
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Folder'
    post:
      tags:
        - Organization
      summary: Crear una nueva carpeta
      operationId: crearCarpeta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FolderCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Folder'

  /folders/{id}:
    delete:
      tags:
        - Organization
      summary: Eliminar una carpeta
      operationId: eliminarCarpeta
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Eliminado.

components:
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          tokenUrl: /auth/token
          scopes: {}
  schemas:
    Entry:
      type: object
      required:
        - titulo
        - type
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        titulo:
          type: string
          example: "Reunión de Prácticas"
        contenido:
          type: string
          description: "Rich text content or encrypted blob"
          example: "Discutir el avance del proyecto SDD."
        type:
          type: string
          enum: [note, task]
        is_encrypted:
          type: boolean
          default: false
        is_deleted:
          type: boolean
          default: false
        encryption_metadata:
          type: string
          description: "JSON string containing client-side encryption metadata (IV, salt, algorithm hints)"
        fecha_creacion:
          type: string
          format: date-time
          readOnly: true
        fecha_modificacion:
          type: string
          format: date-time
          readOnly: true
        version:
          type: integer
          default: 1
          readOnly: true
        tags:
          type: array
          items:
            type: string
        folder_id:
          type: integer
        # Kanban/Task specific fields (only for type=task)
        priority:
          type: string
          enum: [low, medium, high]
        status_column:
          type: string
        due_date:
          type: string
          format: date-time
    
    EntryUpdate:
      type: object
      properties:
        titulo:
          type: string
        contenido:
          type: string
        is_encrypted:
          type: boolean
        is_deleted:
          type: boolean
        encryption_metadata:
          type: string
        tags:
          type: array
          items:
            type: string
        folder_id:
          type: integer
        priority:
          type: string
          enum: [low, medium, high]
        status_column:
          type: string
        due_date:
          type: string
          format: date-time

    KanbanColumn:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        position:
          type: integer
    
    KanbanColumnCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        position:
          type: integer

    Asset:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        content_type:
          type: string
        size:
          type: integer
        entry_id:
          type: integer
          description: "ID de la entrada a la que pertenece este recurso"

    Folder:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        parent_id:
          type: integer
        is_encrypted:
          type: boolean
          default: false
        encryption_metadata:
          type: string
    
    FolderCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        parent_id:
          type: integer
        is_encrypted:
          type: boolean
          default: false
        encryption_metadata:
          type: string

    UserCreate:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    Token:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
    Error:
      type: object
      required:
        - codigo
        - mensaje
      properties:
        codigo:
          type: string
          example: "NOT_FOUND"
        mensaje:
          type: string
          example: "El recurso solicitado no pudo ser localizado."